// src/app/core/services/models/user-profile-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserStatsResponse } from '../models/user-stats-response';
export interface UserProfileResponse {
  dateOfBirth?: string;
  email?: string;
  firstname?: string;
  gender?: 'MALE' | 'FEMALE';
  id?: number;
  lastname?: string;
  online?: boolean;
  skill?: 'COMMUNICATION' | 'ORGANIZATION' | 'IT' | 'TEACHING' | 'FIRST_AID' | 'LOGISTICS' | 'PSYCHOLOGICAL_SUPPORT' | 'LEADERSHIP';
  stats?: UserStatsResponse;
}